Files belonging to SBT "answers" project moved into package fpinscala.answers.  Imports in these files also changed to fpinscala.answers.*
