+++ answers/src/main/scala/fpinscala/datastructures/Tree.scala
@@ -1,4 +1,4 @@
-package fpinscala.datastructures
+package fpinscala.answers.datastructures
 
 sealed trait Tree[+A]
 case class Leaf[A](value: A) extends Tree[A]
