+++ answers/src/main/scala/fpinscala/errorhandling/Either.scala
@@ -1,4 +1,4 @@
-package fpinscala.errorhandling
+package fpinscala.answers.errorhandling
 
 
 //hide std library `Option` and `Either`, since we are writing our own in this chapter
