+++ answers/src/main/scala/fpinscala/iomonad/IO.scala
@@ -1,4 +1,4 @@
-package fpinscala.iomonad
+package fpinscala.answers.iomonad
 
 object IO0 {
                             /*
@@ -240,7 +240,7 @@ object IO2b {
 
 object IO2c {
 
-  import fpinscala.parallelism.Nonblocking._
+  import fpinscala.answers.parallelism.Nonblocking._
 
   /*
    * We've solved our first problem of ensuring stack safety, but we're still
@@ -346,7 +346,7 @@ object IO3 {
   only console I/O effects.
   */
 
-  import fpinscala.parallelism.Nonblocking.Par
+  import fpinscala.answers.parallelism.Nonblocking.Par
 
   sealed trait Console[A] {
     def toPar: Par[A]
