+++ answers/src/main/scala/fpinscala/iomonad/Monad.scala
@@ -1,5 +1,5 @@
-package fpinscala.iomonad
-
+package fpinscala.answers.iomonad
+import scala.language.implicitConversions
 import language.higherKinds // Disable warnings for type constructor polymorphism
 
 trait Functor[F[_]] {
