+++ answers/src/main/scala/fpinscala/iomonad/Throw.scala
@@ -1,4 +1,4 @@
-package fpinscala.iomonad
+package fpinscala.answers.iomonad
 
 /**
  * A version of `TailRec` implemented using exceptions.
