+++ answers/src/main/scala/fpinscala/iomonad/package.scala
@@ -1,7 +1,10 @@
-package fpinscala
+package fpinscala.answers
+import scala.language.implicitConversions
+import scala.language.postfixOps
+import scala.language.higherKinds
 
 package object iomonad {
-  import fpinscala.parallelism.Nonblocking._
+  import fpinscala.answers.parallelism.Nonblocking._
 
   type IO[A] = IO3.IO[A]
   def IO[A](a: => A): IO[A] = IO3.IO[A](a)
