+++ answers/src/main/scala/fpinscala/monoids/Monoid.scala
@@ -1,7 +1,10 @@
-package fpinscala.monoids
+package fpinscala.answers.monoids
 
-import fpinscala.parallelism.Nonblocking._
-import fpinscala.parallelism.Nonblocking.Par.toParOps // infix syntax for `Par.map`, `Par.flatMap`, etc
+import fpinscala.answers.parallelism.Nonblocking._
+import fpinscala.answers.parallelism.Nonblocking.Par.toParOps // infix syntax for `Par.map`, `Par.flatMap`, etc
+
+import fpinscala.answers.parallelism.Nonblocking._
+import fpinscala.answers.parallelism.Nonblocking.Par.toParOps // infix syntax for `Par.map`, `Par.flatMap`, etc
 
 trait Monoid[A] {
   def op(a1: A, a2: A): A
@@ -69,7 +72,7 @@ object Monoid {
     val zero = (a: A) => a
   }
 
-  import fpinscala.testing._
+  import fpinscala.answers.testing._
   import Prop._
 
   def monoidLaws[A](m: Monoid[A], gen: Gen[A]): Prop =
