+++ answers/src/main/scala/fpinscala/parsing/Parsers.scala
@@ -1,9 +1,14 @@
-package fpinscala.parsing
+package fpinscala.answers.parsing
+
+import scala.language.implicitConversions
+// import scala.language.postfixOps
+import scala.language.higherKinds
+
 
 import java.util.regex._
 import scala.util.matching.Regex
-import fpinscala.testing._
-import fpinscala.testing.Prop._
+import fpinscala.answers.testing._
+import fpinscala.answers.testing.Prop._
 
 trait Parsers[Parser[+_]] { self => // so inner classes may call methods of trait
   def run[A](p: Parser[A])(input: String): Either[ParseError,A]
